+++ scout/server/blueprints/variants/controllers.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 import datetime
 import logging
 import os.path
@@ -6,7 +5,7 @@ import urllib.parse
 from datetime import date
 from pprint import pprint as pp
 
-from flask import Response, flash, request, url_for
+from flask import Response, flash, request, url_for, session
 from flask_login import current_user
 from flask_mail import Message
 from werkzeug.datastructures import Headers, MultiDict
@@ -485,7 +484,13 @@ def cancer_variants(store, institute_id, case_name, variants_query, form, page=1
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
     per_page = 50
     skip_count = per_page * max(page - 1, 0)
+
     variant_count = variants_query.count()
+
+    # Setup variant count session with variant count by category
+    variant_count_session(store, institute_id, case_obj["_id"], "clinical", "cancer")
+    session["filtered_variants"] = variant_count
+
     more_variants = True if variant_count > (skip_count + per_page) else False
     variant_res = variants_query.skip(skip_count).limit(per_page)
     data = dict(
@@ -502,6 +507,34 @@ def cancer_variants(store, institute_id, case_name, variants_query, form, page=1
         form=form,
     )
     return data
+
+
+def variant_count_session(store, institute_id, case_id, var_type, var_category):
+    """Create a session object containing variant count for each variant category
+
+    Args:
+        store(scout.adapter.MongoAdapter)
+        institute_id(str): Institute ID
+        case_id(str): Case ID
+        var_type(str): "research" or "clinical"
+        var_category(str): "snv", "cancer", "cancer_sv", "sv"
+    """
+    # Run the query just once for a case
+    if (
+        session.get("institute") != institute_id
+        or session.get("case") != case_id
+        or session["case_variants"] is None
+    ):
+        session["case"] = case_id
+        session["institute"] = institute_id
+        case_variants = store.case_variants_count(case_id, institute_id)
+        session["case_variants"] = case_variants
+
+    if session["case_variants"].get(var_type):
+        session["all_variants"] = session["case_variants"][var_type].get(var_category, "NA")
+    else:
+        # Something might always go wrong I guess?
+        session["all_variants"] = "NA"
 
 
 def get_clinvar_submission(store, institute_id, case_name, variant_id, submission_id):